passport-ocr/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

easyocr==1.7.1
fastapi==0.111.0
matplotlib==3.9.0
mrz==0.6.2
numpy==1.26.2
opencv-python==4.9.0.80
pandas==2.2.2
PassportEye==2.2.1
Pillow==10.1.0
python-dateutil==2.8.2
uvicorn==0.29.0
